fix(docker): clean up Dockerfile and adjust MAX_RMS_AMPLITUDE in config
This commit removes a duplicated CMD instruction in the Dockerfile and also includes clean-up of the commented-out build and tag instructions. Additionally, the MAX_RMS_AMPLITUDE setting within config.json has been decremented to 0.01 from 0.03 to potentially fine-tune the application's audio processing sensitivity.

The docker_run.sh script was updated to explicitly name the running container as 'hey-sheldon-backend' for easier reference and management.