feat(sound-detection): log sox command output on process exit
Added console logging for the standard error stream data accumulated from the 'sox' command upon process exit. This is marked with a FIXME comment to indicate that it is potentially temporary or for debugging purposes and should be reviewed for appropriateness in production code. Previously, the stderr data were collected but not logged, which could make it difficult to diagnose issues with the sound detection process. Logging this information will assist developers in understanding the behavior of the subprocess and facilitate faster debugging and resolution of any issues identified by these logs.